테스트 코드 에러 수정

+++ teamKMS/src/test/java/com/devworker/kms/repo/board/QnaCodeRepoTest.java
@@ -31,7 +31,7 @@ public class QnaCodeRepoTest {
 	MenuRepo menuRepo;
 
 	static QnaCodeDao qnaCodeDao;
-	static int boardId = 1;
+	static long boardId = 1;
 	static int menuId = 4;
 
 	@Before
